+++ R/client_generics.R
@@ -19,11 +19,14 @@
 #'
 #' Storage object management methods:
 #' - `list_storage_files` dispatches to `list_blobs`, `list_azure_files` or `list_adls_files`
-#' - `create_storage_dir` dispatches to `create_azure_dir` or `create_adls_dir`; throws an error if passed a blob container
-#' - `delete_storage_dir` dispatches to `delete_azure_dir` or `delete_adls_dir`; throws an error if passed a blob container
+#' - `create_storage_dir` dispatches to `create_blob_dir`, `create_azure_dir` or `create_adls_dir`
+#' - `delete_storage_dir` dispatches to `delete_blob_dir`, `delete_azure_dir` or `delete_adls_dir`
 #' - `delete_storage_file` dispatches to `delete_blob`, `delete_azure_file` or `delete_adls_file`
 #' - `storage_file_exists` dispatches to `blob_exists`, `azure_file_exists` or `adls_file_exists`
 #' - `storage_dir_exists` dispatches to `blob_dir_exists`, `azure_dir_exists` or `adls_dir_exists`
+#' - `create_storage_snapshot` dispatches to `create_blob_snapshot`
+#' - `list_storage_snapshots` dispatches to `list_blob_snapshots`
+#' - `delete_storage_snapshot` dispatches to `delete_blob_snapshot`
 #'
 #' @seealso
 #' [storage_endpoint], [blob_container], [file_share], [adls_filesystem]
@@ -328,3 +331,36 @@ azure_dir_exists(container, dir, ...)
 #' @export
 storage_dir_exists.adls_filesystem <- function(container, dir, ...)
 adls_dir_exists(container, dir, ...)
+
+
+# snapshots
+
+#' @rdname generics
+#' @export
+create_storage_snapshot <- function(container, file, ...)
+UseMethod("create_storage_snapshot")
+
+#' @rdname generics
+#' @export
+create_storage_snapshot.blob_container <- function(container, file, ...)
+create_blob_snapshot(container, file, ...)
+
+#' @rdname generics
+#' @export
+list_storage_snapshots <- function(container, ...)
+UseMethod("list_storage_snapshots")
+
+#' @rdname generics
+#' @export
+list_storage_snapshots.blob_container <- function(container, ...)
+list_blob_snapshots(container, ...)
+
+#' @rdname generics
+#' @export
+delete_storage_snapshot <- function(container, file, ...)
+UseMethod("delete_storage_snapshot")
+
+#' @rdname generics
+#' @export
+delete_storage_snapshot.blob_container <- function(container, file, ...)
+delete_blob_snapshot(container, file, ...)